разбил код на модули(предварительно)
